requirements.txt
fastapi==0.100.0
sqlmodel==0.0.8
asyncpg==0.28.0
uvicorn==0.22.0 #Asynchronous Server Gateway Interface
httpx==0.23.0
pyyaml
